+++ project-service/src/main/java/com/hps/projectservice/repositories/PhaseRepository.java
@@ -3,6 +3,6 @@ package com.hps.projectservice.repositories;
 import com.hps.projectservice.entities.Phase;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PhaseRepository extends JpaRepository<Phase, Integer> {
+public interface PhaseRepository extends JpaRepository<Phase, Long> {
     public Phase findPhaseByName(String name);
 }
