+++ project-service/src/main/java/com/hps/projectservice/services/interfaces/PhaseService.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public interface PhaseService {
     public List<Phase> getAllPhases();
-    public Phase getPhaseById(Integer id);
+    public Phase getPhaseById(Long id);
     public Phase getPhaseByName(String name);
     public Phase createNewPhase(Phase phase);
-    public Phase updatePhase(Integer id, Phase phase);
-    public void deletePhaseById(Integer id);
+    public Phase updatePhase(Long id, Phase phase);
+    public void deletePhaseById(Long id);
 }
